JSON decoding looks for doubles now, not ints.

+++ client/JsonEncodeDecode.cpp
@@ -75,7 +75,7 @@ GameState* JsonEncodeDecode::decode(char* buffer)
 int JsonEncodeDecode::decodeVehicle(GameState* state, const Value& vehicle)
 {
 	// Vehicle vars
-	int x = 0, y = 0, velocity = 0, angle = 0, endurance = 0;
+	double x = 0, y = 0, velocity = 0, angle = 0, endurance = 0;
 
 	// Iterate through JSON vehicle object
 	const char* check;
@@ -85,33 +85,33 @@ int JsonEncodeDecode::decodeVehicle(GameState* state, const Value& vehicle)
 
 		if (strcmp(check, "X") == 0)
 		{
-			x = itr->value.GetInt();
+			x = itr->value.GetDouble();
 			continue;
 		}
 		if (strcmp(check, "Y") == 0)
 		{
-			y = itr->value.GetInt();
+			y = itr->value.GetDouble();
 			continue;
 		}
 		if (strcmp(check, "Velocity") == 0)
 		{
-			velocity = itr->value.GetInt();
+			velocity = itr->value.GetDouble();
 			continue;
 		}
 		if (strcmp(check, "Angle") == 0)
 		{
-			angle = itr->value.GetInt();
+			angle = itr->value.GetDouble();
 			continue;
 		}
 		if (strcmp(check, "Endurance") == 0)
 		{
-			endurance = itr->value.GetInt();
+			endurance = itr->value.GetDouble();
 			continue;
 		}
 	}
 
 	// Create a new Vehicle
-	Vehicle* new_vehicle = new Vehicle(x, y, endurance, angle, velocity);
+	Vehicle* new_vehicle = new Vehicle((int) x, (int) y, (int) endurance, angle, velocity);
 
 	// Push it to the GameState
 	state->addPlayer(new_vehicle);
